fix: the sign up CTA should only be shown if not signed in
